requirements.txt
asgiref==3.8.1
cryptography==39.0.2
Django==4.2.17
django-allauth==65.3.1
django-cors-headers==3.12.0
django-environ==0.11.2
django-rest-auth==0.9.5
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==24.2
PyJWT==2.8.0
six==1.17.0
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata
requests
dj-rest-auth
#KoNLPy>=0.6.0     
#JPype1>=1.4.1    
#spacy>=3.7.2 
